Extract `IntervalModel` to `pycroft.model`
This keeps the `helpers.interval` module free from sqla-imports,
which is important bebcause I ran into circular imports.

It also makes sense because the main purpose of that module is to
introduce a python datatype, which is completely unrelated to
databases.

Refs #160
